Initial cut at exposing settings for use out of `AutoPlugin` context for more advanced config scoping.
