+++ src/main/scala/com/typesafe/sbt/site/SiteScaladocPlugin.scala
@@ -1,7 +1,7 @@
-package com.typesafe.sbt.site.util
+package com.typesafe.sbt.site
 
 import com.typesafe.sbt.site.SitePlugin.autoImport.siteSubdirName
-import com.typesafe.sbt.site.SitePlugin
+import com.typesafe.sbt.site.util.SiteHelpers
 import sbt.Keys._
 import sbt._
