+++ src/main/scala/com/typesafe/sbt/site/jekyll/JekyllPlugin.scala
@@ -2,42 +2,42 @@ package com.typesafe.sbt.site.jekyll
 
 import com.typesafe.sbt.site.SitePlugin.autoImport.siteSubdirName
 import com.typesafe.sbt.site.SitePlugin
-import com.typesafe.sbt.site.util.SiteHelpers
+import com.typesafe.sbt.site.util.RubyHelpers.RubyKeys
+import com.typesafe.sbt.site.util.{RubyHelpers, SiteHelpers}
 import sbt.Keys._
 import sbt._
+
+/** Support to generate a Jekyll-base website. */
 object JekyllPlugin extends AutoPlugin {
   override def requires = SitePlugin
   override def trigger = noTrigger
 
-  object autoImport {
+  object autoImport extends RubyKeys {
     val Jekyll = config("jekyll")
-    val requiredGems = SettingKey[Map[String, String]](
-      "jekyll-required-gems", "Required gem + versions for this build.")
-    val checkGems = TaskKey[Unit](
-      "jekyll-check-gems", "Tests whether or not all required gems are available.")
   }
   import autoImport._
-  override def projectSettings: Seq[Setting[_]] =
-    SiteHelpers.directorySettings(Jekyll) ++
-      Seq(
-        includeFilter in Jekyll := ("*.html" | "*.png" | "*.js" | "*.css" | "*.gif" | "CNAME" | ".nojekyll"),
-        requiredGems := Map.empty
-      ) ++
-      inConfig(Jekyll)(
-        Seq(
-          checkGems := SiteHelpers.checkGems(requiredGems.value, streams.value),
-          mappings := {
-            val cg = checkGems.value
-            generate(sourceDirectory.value, target.value, includeFilter.value, streams.value)
-          },
-          siteSubdirName := ""
-        )
-      ) ++
-      SiteHelpers.watchSettings(Jekyll) ++
-      SiteHelpers.addMappingsToSiteDir(mappings in Jekyll, siteSubdirName in Jekyll)
+  override def projectSettings = jekyllSettings(Jekyll)
 
-  // TODO - Add command line args and the like.
-  final def generate(
+  /** Creates the settings necessary for running Jekyll in the given configuration. */
+  def jekyllSettings(config: Configuration): Seq[Setting[_]] =
+    inConfig(config)(
+      Seq(
+        checkGems := RubyHelpers.checkGems(requiredGems.value, streams.value),
+        requiredGems := Map.empty,
+        includeFilter := ("*.html" | "*.png" | "*.js" | "*.css" | "*.gif" | "CNAME" | ".nojekyll"),
+        mappings := {
+          val _ = checkGems.value
+          generate(sourceDirectory.value, target.value, includeFilter.value, streams.value)
+        },
+        siteSubdirName := ""
+      )
+    ) ++
+      SiteHelpers.directorySettings(config) ++
+      SiteHelpers.watchSettings(config) ++
+      SiteHelpers.addMappingsToSiteDir(mappings in config, siteSubdirName in config)
+
+  /** Run jekyll via fork. TODO - Add command line args and the like. */
+  private[sbt] def generate(
     src: File,
     target: File,
     inc: FileFilter,
@@ -52,4 +52,6 @@ object JekyllPlugin extends AutoPlugin {
       (file, name) <- target ** inc --- target pair relativeTo(target)
     } yield file -> name
   }
+
+
 }
