+++ src/main/scala/com/typesafe/sbt/site/nanoc/NanocPlugin.scala
@@ -3,45 +3,41 @@ package com.typesafe.sbt.site.nanoc
 import java.io.FileReader
 import com.typesafe.sbt.site.SitePlugin.autoImport.siteSubdirName
 import com.typesafe.sbt.site.SitePlugin
-import com.typesafe.sbt.site.util.SiteHelpers
+import com.typesafe.sbt.site.util.RubyHelpers.RubyKeys
+import com.typesafe.sbt.site.util.{RubyHelpers, SiteHelpers}
 import sbt.Keys._
 import sbt._
 
 import scala.collection.immutable
-/** Nanoc generator. */
+/** Support ot generato a Nanoc-based website. */
 object NanocPlugin extends AutoPlugin {
   override def requires = SitePlugin
   override def trigger = noTrigger
-  object autoImport {
+  object autoImport extends RubyKeys {
     val Nanoc = config("nanoc")
-
-    val requiredGems = SettingKey[Map[String, String]](
-      "nanoc-required-gems", "Required gem + versions for this build.")
-    val checkGems = TaskKey[Unit](
-      "nanoc-check-gems", "Tests whether or not all required gems are available.")
   }
-  import autoImport._
-  override def projectSettings: Seq[Setting[_]] =
-    SiteHelpers.directorySettings(Nanoc) ++
-      Seq(
-        includeFilter in Nanoc := AllPassFilter,
-        requiredGems := Map.empty
-      ) ++
-      inConfig(Nanoc)(
-        Seq(
-          checkGems := SiteHelpers.checkGems(requiredGems.value, streams.value),
-          mappings := {
-            val cg = checkGems.value
-            generate(
-              sourceDirectory.value, target.value, includeFilter.value, streams.value)
-          },
-          siteSubdirName := ""
-        )
-      ) ++
-      SiteHelpers.watchSettings(Nanoc) ++
-      SiteHelpers.addMappingsToSiteDir(mappings in Nanoc, siteSubdirName in Nanoc)
 
-  // TODO - Add command line args and the like.
+  import autoImport._
+  override def projectSettings = nanocSettings(Nanoc)
+  def nanocSettings(config: Configuration): Seq[Setting[_]] =
+    inConfig(config)(
+      Seq(
+        requiredGems := Map.empty,
+        checkGems := RubyHelpers.checkGems(requiredGems.value, streams.value),
+        includeFilter := AllPassFilter,
+        mappings := {
+          val cg = checkGems.value
+          generate(
+            sourceDirectory.value, target.value, includeFilter.value, streams.value)
+        },
+        siteSubdirName := ""
+      )
+    ) ++
+      SiteHelpers.directorySettings(config) ++
+      SiteHelpers.watchSettings(config) ++
+      SiteHelpers.addMappingsToSiteDir(mappings in config, siteSubdirName in config)
+
+  /** Run nanoc via fork. TODO - Add command line args and the like. */
   private[sbt] def generate(
     src: File,
     target: File,
